Теперь модули возвращают функции
