+++ js/main.js
@@ -1,2 +1,7 @@
-import './avatar-render.js';
+import {getPhotos} from './temp-data.js';
+import {createPictures} from './pictures.js';
+
+const photos = getPhotos();
+createPictures(photos);
+
